requirements.txt
apify-client==1.7.2
python-dotenv==1.0.0
pandas==2.1.4
